requirements.txt
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.29.0
matplotlib>=3.7.0
tensorboard>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.4.0
